==> teste-progresso/06-ml-comparison/07-clusters-analysing-differences.ipynb <==
import pandas as pd

annotation_sets = ['Human_84', 'BioBERT_Llama']

metrics_all = {'Human_84': pd.read_csv('medical_specialist/annotations-dpoc-medical_specialist_metrics_biobert.csv'),
               'BioBERT_Llama': pd.read_csv('biobert-llama_balanced/annotations-dpoc-biobert-llama_metrics.csv')}

merged_df = pd.merge(metrics_all['Human_84'], metrics_all['BioBERT_Llama'], on='annotation id', suffixes=('_human', '_biobert'))
filtered_df = merged_df[(merged_df['cluster kmeans 3 cat_human'] == 2) & (merged_df['cluster kmeans 3 cat_biobert'] == 0)]

filtered_df
